+++ Flow.Launcher.Plugin/IContextMenu.cs
@@ -1,0 +1,9 @@
+﻿using System.Collections.Generic;
+
+namespace Flow.Launcher.Plugin
+{
+    public interface IContextMenu : IFeatures
+    {
+        List<Result> LoadContextMenus(Result selectedResult);
+    }
+}
