+++ Flow.Launcher.Plugin/IPluginI18n.cs
@@ -1,0 +1,12 @@
+﻿namespace Flow.Launcher.Plugin
+{
+    /// <summary>
+    /// Represent plugins that support internationalization
+    /// </summary>
+    public interface IPluginI18n : IFeatures
+    {
+        string GetTranslatedPluginTitle();
+
+        string GetTranslatedPluginDescription();
+    }
+}
